prepare usage of other CsvWriters (e.g. ZIP-based)

+++ src/ad/cppgtfs/gtfs/Route.h
@@ -36,7 +36,7 @@ class RouteB {
   RouteB(const string& id, typename AgencyT::Ref agency,
          const string& short_name, const string& long_name, const string& desc,
          flat::Route::TYPE type, const string& url, uint32_t color,
-         uint32_t text_color, size_t sort_order)
+         uint32_t text_color, int64_t sort_order)
       : _id(id),
         _agency(agency),
         _short_name(short_name),
@@ -76,7 +76,7 @@ class RouteB {
     return flat::Route::getHexColorString(_text_color);
   }
 
-  size_t getSortOrder() const { return _sort_order; }
+  int64_t getSortOrder() const { return _sort_order; }
 
   flat::Route getFlat() const {
     flat::Route r;
@@ -109,7 +109,7 @@ class RouteB {
   string _url;
   uint32_t _color;
   uint32_t _text_color;
-  size_t _sort_order;
+  int64_t _sort_order;
 };
 
 typedef RouteB<Agency> Route;
